move docs.py to docs/generate.py
